refactor(renderer): split performance monitoring
